requirements.txt
matplotlib
numpy
PyQt6
requests
pymysql
flask-restx